% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_com.R
\name{check_com}
\alias{check_com}
\title{Check Consistency in Community Data}
\usage{
check_com(com, remove_zeros = T)
}
\arguments{
\item{com}{a community data object.}

\item{remove_zeros}{whether to remove samples and species with zero observations.}
}
\value{
A list with community data.
}
\description{
Checks for consistency among datasets within a 'com' list. Removes samples and species that are not present in all datasets, and optionally those with zero observations.
}
\examples{
check_com()
}
\keyword{community}
\keyword{data}
