% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_com.R
\name{diversity_com}
\alias{diversity_com}
\title{Calculates Diversity Indices For Communitiy data}
\usage{
diversity_com(com, file = "diversity_indices.csv", ...)
}
\arguments{
\item{com}{a community data object.}

\item{file}{the name of the file where a table with the diversity values is exported.}

\item{...}{further arguments to be passed to 'write.table'.}
}
\value{
A list with community data.
}
\description{
Calculates abundance and various diversity indices per sample for a community data object. The diversity values are expressed in effective species numbers.
}
\examples{
diversity_com()
}
\keyword{community}
\keyword{data}
