% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_norm.R
\name{mm_norm}
\alias{mm_norm}
\title{Mixture Model Normalization Of Sequence Reads}
\usage{
mm_norm(com, ...)
}
\arguments{
\item{com}{a community data object.}

\item{...}{further arguments to be passed to 'estimateDispersions'.}
}
\value{
A list with community data.
}
\description{
Normalizes sequence reads in community data matrix with a mixture model using DESeq.
}
\examples{
mm_norm()
}
\keyword{community}
\keyword{data}
