% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_taxonomy.R
\name{modify_taxonomy}
\alias{modify_taxonomy}
\title{Modify Taxon Names}
\usage{
modify_taxonomy(x, term = "unclassified")
}
\arguments{
\item{x}{taxonomy table.}

\item{term}{term used for unclassified/unidentified taxa.}
}
\value{
A data frame with taxonomic categories.
}
\description{
Modifies taxon names in a taxonomy table.
}
\details{
Adds to unclassified/unidentified taxa the lowest taxon name known.
}
\examples{
modify_taxonomy()
}
\keyword{taxonomy}
