% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_clustering.R
\name{otu_clustering}
\alias{otu_clustering}
\title{Clusters Species By Sequence Similarity}
\usage{
otu_clustering(
  com,
  thr = 97,
  method = "cd-hit",
  folder = "OTU_clustering",
  cleanup_files = F
)
}
\arguments{
\item{com}{a community data object.}

\item{thr}{threshold of sequence similarity for OTU clustering. In percentage.}

\item{method}{method of clustering, either 'cd-hit' or 'blastclust'}
}
\value{
A list with community data.
}
\description{
Clusters species in a community data object into Operational Taxonomy Units (OTUs) by sequence similariy.
}
\details{
The function still depends of an external bash script ('otu_clustering.sh'), whcih makes portability problematic. Future changes will aim at solving this problem.
}
\examples{
otu_clustering()
}
\keyword{community}
\keyword{data}
