% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_diversity_com.R
\name{phylo_diversity_com}
\alias{phylo_diversity_com}
\title{Phylogenetic Diversity Of Communitiy data}
\usage{
phylo_diversity_com(
  com,
  runs = 100,
  file = "phylo_diversity.RData",
  use_saved = F
)
}
\arguments{
\item{com}{a community data object.}

\item{runs}{number of randomizations.}

\item{file}{the name of the file where the data will be stored as 'RData' object.}

\item{use_saved}{read data from 'file' instead of importing them from scratch.}
}
\value{
A list with community data.
}
\description{
Calculates phylogenetic diversity measures per sample for a community data object.
}
\examples{
phylo_diversity_com()
}
\keyword{community}
\keyword{data}
