% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vp_euler4.R
\name{plot_vp_euler4}
\alias{plot_vp_euler4}
\title{Plot Euler Diagram From Variation Partition Results (4 Vars)}
\usage{
plot_vp_euler4(
  vp,
  names = c("X1", "X2", "X3", "X4"),
  col = c("brown3", "skyblue3", "orange", "chartreuse4")
)
}
\arguments{
\item{vp}{output from 'varpart' analysis.}

\item{vector}{with the names of the tables with explanatory variables.}

\item{color}{vector with colors for the diagram bubbles.}
}
\value{
An Euler plot.
}
\description{
Plots an Euler diagram from an object created with vegan's 'varpart' function, using four explanatory variables.
}
\examples{
plot_vp_euler3()
}
\keyword{community}
\keyword{data}
