% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_com.R
\name{pool_com}
\alias{pool_com}
\title{Pool Community Data By Factor}
\usage{
pool_com(com, var, fun = mean)
}
\arguments{
\item{com}{a community data object.}

\item{var}{the name of a column in 'env' to use for pooling.}

\item{fun}{function to use for pooling. Default is mean.}
}
\value{
A list with community data.
}
\description{
Pools samples in a community data object by a factor in the 'env' table.
}
\examples{
pool_com()
}
\keyword{community}
\keyword{data}
