% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_com.R
\name{preprocess_com}
\alias{preprocess_com}
\title{Preprocess Community Data}
\usage{
preprocess_com(
  com,
  min_obs = 100,
  min_taxa = 5,
  normalize = T,
  method = "blind",
  sharingMode = "maximum",
  fitType = "local",
  ...
)
}
\arguments{
\item{com}{a community data object.}

\item{min_obs}{remove samples with a total number of observations below this treshold.}

\item{min_taxa}{remove species with a total number of observations below this treshold.}

\item{normalize}{logical, whether to normalize abundances using a mixture model with DESeq.}

\item{method}{arguments to be passed to 'estimateDispersions'.}

\item{sharingMode}{arguments to be passed to 'estimateDispersions'.}

\item{fitType}{arguments to be passed to 'estimateDispersions'.}

\item{...}{further arguments to be passed to 'estimateDispersions'.}
}
\value{
A list with community data.
}
\description{
Performs preprocessing tasks in community data, removing samples and species with few observations, and normalizing abundance data.
}
\examples{
preprocess_com.com()
}
\keyword{community}
\keyword{data}
