% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_com.R
\name{read_com}
\alias{read_com}
\title{Community Data Input}
\usage{
read_com(
  cdm,
  env,
  tax,
  seq = NULL,
  phy = NULL,
  file = NULL,
  use_saved = F,
  ...
)
}
\arguments{
\item{cdm}{the name of the file with the community data matrix. Rows are samples, and columns are species.}

\item{env}{the name of the file with a table with environmental data. First column must include sample names as in rows of 'cdm'.}

\item{tax}{the name of the file with a table with taxonomic information. First column must include species names as in columns of 'cdm'.}

\item{seq}{the name of the file where sequences for each species in fasta format. Sequence headers must be the species names as in columns of 'cdm'.}

\item{phy}{the name of the file with a phylogenetic tree in Newick format including the species in 'cdm'. Tip labels must be the species names as in columns of 'cdm'. If no tree is provided, a taxonomic tree can be automatically generated using the information in 'tax', using option 'tree_from_tax = T'.}

\item{file}{the name of the file where the data will be stored as 'RData' object.}

\item{use_saved}{read data from 'file' instead of importing them from scratch.}

\item{...}{further arguments to be passed to 'read.table'.}
}
\value{
A list with community data.
}
\description{
Reads community data from files.
}
\details{
If no phylogenetic tree is provided, the calculation from the taxonomic data provided is performed using the perl script 'taxonomy_to_tree.pl' available at 'https://github.com/mr-y/my_bioinfPerls'.
}
\examples{
read_com()
}
\keyword{community}
\keyword{data}
