% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_data.R
\name{soil_data}
\alias{soil_data}
\title{Add Soil Data To Community Data Object}
\usage{
soil_data(com, soil, eigenvectors = T, k = 3)
}
\arguments{
\item{com}{a community data object.}

\item{soil}{data frame with soil charactersitics to be added. Row names must be as in 'env'.}

\item{eigenvectors}{logical, whether eigenvectors should be provided instead of raw soil data.}

\item{k}{the number of eigenvectors retained.}
}
\value{
A list with community data.
}
\description{
Adds soil characteristics to community data object.
}
\details{
If eigenvectors = T, eigenvalues are obtained from a Principal Coordinates Analysis calculated from the soil data.
}
\examples{
soil_data()
}
\keyword{community}
\keyword{data}
