% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data.R
\name{spatial_data}
\alias{spatial_data}
\title{Add Geographic Coordinates To Community Data}
\usage{
spatial_data(com, x, y)
}
\arguments{
\item{com}{a community data object.}

\item{x}{name of the column in 'env' where the longitude coordinates are stored.}

\item{y}{name of the column in 'env' where the latitude coordinates are stored.}
}
\value{
A list with community data.
}
\description{
Adds grographic coordinates as a separate table in a community data object.
}
\examples{
spatial_data()
}
\keyword{community}
\keyword{data}
