% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_prop.R
\name{tax_prop}
\alias{tax_prop}
\title{Proportion of Taxa In Community Data}
\usage{
tax_prop(
  com,
  taxon,
  var = NULL,
  n = length(levels(com$tax[, taxon])),
  fun = sum,
  omit_unclassified = T,
  term = "unclassified"
)
}
\arguments{
\item{com}{a community data object.}

\item{taxon}{name of the taxonomic level to calculate proportions. Must be a column name in table 'tax'.}

\item{var}{a factor to pool samples. Must be a column name in table 'env'. Optional.}

\item{fun}{function to use for pooling taxonomic data. Default is sum.}

\item{omit_unclassified}{omit unclassified taxa from the column sorting. Unclassified taxa are placed at the end.}

\item{term}{if omit_unclassified = T, the term defining unclassified taxa.}
}
\value{
A data frame with proportions of taxa across samples.
}
\description{
Calculates the proportion of taxonomic categories across samples of a community data object.
}
\examples{
tax_prop()
}
\keyword{community}
\keyword{data}
