% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxtotree.R
\name{taxtotree}
\alias{taxtotree}
\title{Taxonomy Table To Tree}
\usage{
taxtotree(tax)
}
\arguments{
\item{tax}{a taxonomy data frame.}
}
\value{
A phylogenetic tree in 'ape' format.
}
\description{
Generates a tree from the taxonomy table in a 'com' object.
}
\examples{
taxtotree()
}
\keyword{community}
\keyword{data}
