% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vp3.R
\name{test_vp3}
\alias{test_vp3}
\title{Significance Testing For Variation Partition (3 Vars)}
\usage{
test_vp3(vp, cdm = NULL)
}
\arguments{
\item{vp}{output from 'varpart' analysis.}
}
\value{
A data frame containing the variance explained and p-values per explanatory variable.
}
\description{
Calculates significance of components of an object created with vegan's 'varpart' function, using three explanatory variables.
}
\details{
If the response data is a distance matrix, calculates p-values from db-RDA analyses. If it is a table, uses RDA.#'
}
\examples{
test_vp3()
}
\keyword{community}
\keyword{data}
