% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_vp4.R
\name{test_vp4}
\alias{test_vp4}
\title{Significance Testing For Variation Partition (4 Vars)}
\usage{
test_vp4(vp)
}
\arguments{
\item{vp}{output from 'varpart' analysis.}
}
\value{
A data frame containing the variance explained and p-values per explanatory variable.
}
\description{
Calculates significance of components of an object created with vegan's 'varpart' function, using four explanatory variables.
}
\details{
If the response data is a distance matrix, calculates p-values from db-RDA analyses. If it is a table, uses RDA.#'
}
\examples{
test_vp4()
}
\keyword{community}
\keyword{data}
