% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_data.R
\name{clim_data}
\alias{clim_data}
\title{Get Bioclimatic Data For Community Data}
\usage{
clim_data(com, x = "lon", y = "lat", k = 3, keep_data = T)
}
\arguments{
\item{com}{a community data object.}

\item{x}{name of the column in 'env' where the longitude coordinates are stored.}

\item{y}{name of the column in 'env' where the latitude coordinates are stored.}

\item{k}{the number of eigenvectors retained.}

\item{keep_data}{logical, whether the downloaded bioclimatic database should be kept after extracting the data.}
}
\value{
A list with community data.
}
\description{
Obtains bioclimatic variables from the "WorldClim" dataset for the samples in a community data object, using geographical coordinates provided in the environmental table.
}
\details{
Eigenvectors are obtained from a Principal Coordinates Analysis calculated from the bioclimatic data.
}
\examples{
clim_data()
}
\keyword{community}
\keyword{data}
